requirements.txt
aiohttp==3.8.4
aiomisc==17.2.2
aiosignal==1.3.1
astroid==2.15.4
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
colorlog==6.7.0
dill==0.3.6
exceptiongroup==1.1.1
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
logging-journald==0.6.4
mccabe==0.7.0
multidict==6.0.4
packaging==23.1
platformdirs==3.5.0
pluggy==1.0.0
pylint==2.17.3
pytest==7.3.1
python-dotenv==1.0.0
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
wrapt==1.15.0
yarl==1.9.2
